Inline custom-element
